requirements.txt
fastapi[standard]
python-dotenv
pymongo
passlib[bcrypt]
uvicorn